Better ensures ProjectCollection is configured ad propert times. Still deadlocks
This commit is better in the sense that it tries to configure the ProjectCollection after full load of the solution.  Also it find the ProjectCollection 1) searching all the DTE loaded projects 2) basing on the project being loaded. It seems more correct but now it triggers deadlock even on solution load of a C++ project: it seems the deadlock is avoidable only if the project is not fully configured. awit code seems correct