Fixed VCProject not dirty when setting global properties
Added a little hack to ensure projects are really marked dirty when setting
the global properties SolutionConfiguration, SolutionPlatform. Before
the Microsoft.Build.Evalutation.Project was marked dirty but the VCProject
was still not marked dirty or considered up-to date anyway. Adding a
dummy property and removing it immediately is enough to ensure
the VCProject is marked is no more up-to date